requirements.txt
streamlit
pandas
yfinance
plotly
numpy
huggingface_hub
os
python-dotenv